convert applayout component to typescript
