+++ src/lib/component/base/AvatarIcon/index.tsx
@@ -1,1 +1,2 @@
+export * from "./AvatarIcon";
 export { default } from "./AvatarIcon";
