+++ src/lib/component/base/index.tsx
@@ -30,6 +30,7 @@ export {
   DictionaryEntry,
 } from "./Dictionary";
 export { default as Duration } from "./Duration";
+export { default as Drawer } from "./Drawer";
 export { default as ErrorDialog } from "./ErrorDialog";
 export { default as FilterList } from "./FilterList";
 export { default as InlineLink } from "./InlineLink";
