+++ src/lib/component/partial/AppLayout/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./AppLayout";
