+++ src/lib/component/view/EntitiesView.tsx
@@ -118,14 +118,14 @@ export const EntitiesViewContent = ({
 
   if (!data.namespace && !loading && !aborted) {
     return (
-      <AppLayout namespace={variables.namespace}>
+      <AppLayout>
         <NotFound />
       </AppLayout>
     );
   }
 
   return (
-    <AppLayout namespace={variables.namespace}>
+    <AppLayout>
       <div>
         <Content marginBottom>
           <EntitiesListToolbar
