+++ src/lib/component/view/EventsView.tsx
@@ -134,14 +134,14 @@ export const EventsViewContent = ({
 
   if (!data.namespace && !loading && !aborted) {
     return (
-      <AppLayout namespace={variables.namespace}>
+      <AppLayout>
         <NotFound />
       </AppLayout>
     );
   }
 
   return (
-    <AppLayout namespace={variables.namespace}>
+    <AppLayout>
       <div>
         <Content marginBottom>
           <EventsListToolbar
