+++ src/lib/component/view/HandlersView.tsx
@@ -121,14 +121,14 @@ export const HandlersViewContent = ({
 
   if (!data.namespace && !loading && !aborted) {
     return (
-      <AppLayout namespace={variables.namespace}>
+      <AppLayout>
         <NotFound />
       </AppLayout>
     );
   }
 
   return (
-    <AppLayout namespace={variables.namespace}>
+    <AppLayout>
       <div>
         <Content marginBottom>
           <HandlersListToolbar
