+++ src/lib/component/view/MutatorsView.tsx
@@ -118,14 +118,14 @@ export const MutatorsViewContent = ({
 
   if (!data.namespace && !loading && !aborted) {
     return (
-      <AppLayout namespace={variables.namespace}>
+      <AppLayout>
         <NotFound />
       </AppLayout>
     );
   }
 
   return (
-    <AppLayout namespace={variables.namespace}>
+    <AppLayout>
       <div>
         <Content marginBottom>
           <MutatorsListToolbar
